requirements.txt
streamlit>=1.31.0
playwright>=1.44.0
groq>=0.4.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openai-whisper>=20231117
fpdf2>=2.7.8
pydub>=0.25.1
SpeechRecognition>=3.10.0
pdfplumber>=0.10.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
toml>=0.10.2
